+++ src/types/rabbitmq-sender-options.ts
@@ -1,9 +1,14 @@
+import type { RoutingKeys } from "src/types/routing-keys";
+
 // Интерфейс опций для настройки RabbitMQModule (отправка сообщений)
 export default interface RabbitMQSenderOptions {
 	// Обязательные параметры для отправки сообщений
 	url: string;
 	exchange: string;
 	exchangeType: string;
+
+	// Обязательный список routing keys для отправки сообщений
+	routingKeys: RoutingKeys;
 
 	// Опциональные параметры для reply очередей RPC паттерна
 	replyQueueOptions?: {